New object: pitch->bands-pitchclass and accompanying example
New object: pitch->bands-pitchclass:
Creates pitches for a set of attacks. Pitches are selected randomly
from within a band that has a trajectory and that is designated by two
bpfs.  Pitches taken from this band are further constrained to a
specified set of pitch classes.

There is also an included example patch called ‘5c
pitch-bands-pitchclass.’