New object Pitch-mapper2
Committing the new object pitch-mapper2.  This object maps random
pitches based on the number of attacks in a rhythm list.  In mode 0
(default) pitches are constrained to a particular pitch collection.  In
mode 1 pitches are constrained by a given pitch range.  In both modes,
the pitch collections can be shared from one pool or from distinct
pools provided for each voice.

An example patch called “5b pitch-mapper2” is also included.