Permutations+Canon-Growth + example
Example patch provided is an the updated “2b-other combination
utilities”

New object: Permutations+Canon-Growth

Provided with a group of lists, iteratively builds output lists with no
rhythmic overlaps, except for the onset.

Builds according to this process: take the first list and check it
against the second list.  If there are no attack overlaps add both
lists to the output list.  Then check the third list agains this output
list. If there are no outputs, combine the third list to the output
list and continue. Etc...

Does this process on each rotation of the original lists provided.

(Other object:Permutations+Canon-Growth-Return-All is in progress)