r-interleave and poly2player updates and example patch
Updated r-interleave in a variety of ways: allows for arbitrary no. of
inputs.  Also, it is now compatible for all forms of poly player
(sustain, rests, or pulsed) (!)

Providing example patch “3e-substitutions with r-interleave 2”

More info on r-interleave:

Interleave an arbitrary number of rhythmic frames
using the rhythmic frame's "voices" output.
Increase the number of inputs in r-interleave object for any number of
rhythmic frames.