Overview patch and revised and renamed example patches
Example patches were renamed to correspond with category (query,
rhythm, pitch, etc…)
Example patches were reorganized and combined together when it made
sense in terms of function.
Example patches were further developed and commented.

An encompassing overview patch was created with short descriptions and
examples for each type of object contained in the library.  The
overview, itself, is also broken down in terms of category for easy
reference.