Q-combi-from-elements object and example patch
new object that runs a special test with q-combi object.
example patch called q-combinations-elements-ex
functionality:
MODE 1. Q-combi-from-elements
Finds combinations of <n> elements which sum up to any one of the
elements provided and provides all rotations of these combinations.
(Ordered/unordered and with or without repetitions.)

MODE 2. Q-combi-from-elements
Finds combinations of <n> elements which sum up to any one of the
elements provided and provides all rotations of these combinations that
meet the q-canon test. (Ordered/unordered and with or without
repetitions.)