Mode 2 for o-sum-lists
Created “mode 2” of o-um-lists. It takes a lists of lists and then
replaces each multi-element list with the sum and also adds a list at
end that sums across left to right.  This is useful for handling the
tatum/beat resolution issue as well as meter frame and it also shows
quickly if there is a calculation error somewhere.