New modes for s-poly and s-poly2 objects and help files
A) In the s-poly object, new modes 2 and 3  allow for changing
harmonies following the rhythmic structure by matching sublists of
pitches and rhythms in a voice.  Mode 2 is the sustain mode.  Mode 3 is
the rests mode. For the s-poly version, please see the new help patch
"s-poly modes 2 and 3" to try this out.

B) In the s-poly2 object, this works the same way.  New mode 3 is for
sustain, new mode 4 is for rests and new mode 5 is for pulse. For the
s-poly2 version, please see the new patch "s-poly2 modes 3 4 and 5."