Various Revisions & New Objects
1) Submitted newly revised example patch “s-poly2 modes 3 4 and 5.”

2) Created object o-list-info that returns (length sum) pairs and
retains the sublist structure of the input list.

The O-list-info object is documented in the revised
“o-operations-ex.om” patch.

3) Updated the o-list-trans object.  It has a new second argument that
allows you to specify the number of transition output lists for the
function to make.

Changes to o-list-trans object are documented in the updated
“o-list-generators” patch

4) Updated q-combi-filter object to include new optional argument
“exclude.”  The exclude argument must be a list.  If any element in the
exclude list appears in a result list then that result list will be
removed from the lists in the q-combi-filter output.

The new exclude argument is demonstrated in the q-combination-utilities
patch.