U-info object and example patch
Created new object u-info. U-info is a utility that provides
information on  voice and poly objects.  The output is a list of pitch
lists, rhythm lists, and number of elements for each voice.  Use the
"self" output of a voice or poly object as input into u-info.

U-info is documented in the new example patch u-info.omp.