O-LIST-REPEAT mode 1 and example
Created O-LIST-REPEAT mode 1.  This takes in a list or list of lists
and a desired length.  It then repeats or truncates a list as necessary
to fill the desired length of elements specified.  The output includes
the tailored list (with repetitions as necessary) and a list of lengths
confirming the desired length.

O-LIST-REPEAT mode 1 is documented in the o-operations-ex-revised2
example patch.