updates to u-score-lists & p-bands-collection
U-score-lists now returns rhythms and tatum lists for voice or poly
objects that are given as input.

U-score-lists documented in u-score-lists.omp example patch

P-bands-collection updated so that it works for multiple pitch
collections.  If one pitch collection is given, the object applies that
collection to all voices.  If more than one pitch collection is given
then the object assumes that there is a collection for each voice and
applies pitch collections to corresponding voices.

Changes to p-bands-collection are documented in the updated p-bands
example patch.