new utilities
These are new utilities for Tessellate.

U-retro
U-retro returns the retrograde of each list in a list of lists
(default).  Optional mode=1 returns the retrograde of each sublist
within the lists of lists.

U-retro is documented in the new u-retro documentation patch.

U-+ is a utility for quick addition of a number to a lists of numbers.
Default mode is normal addition. Optional Mode=1 is mod12-based
addition.  Mode=1 assumes midics as input.  Optional mod=12 allows you
to change the mod number if you don’t want to use the typical mod12
value

u-+ is documented in the new  u-arithmetic patch

U-- is a utility for quick subtraction of a number from a lists of
numbers.  Default mode is normal subtraction. Optional Mode=1 is
mod12-based subtraction. Mode=1 assumes midics as input.  Optional
mod=12 to change mod number.

u— is documented in the new u-arithmetic patch

u-* = as above but for multiplication.

u-inversion Inverts a list of pitches around a provided  pitch.
Mode=1 allows for mod12 and assumes input to be midics.
Optional mod input allows for different mod values.

u-inversion is documented in the new u-arithmetic patch.