pc conversion utilities
"U-pc->midic takes in a list of list of pcs and converts them to lists
of midics using a reference pitch to set the register. The reference
pitch should be a midic, ideally a C pitchclass. The output midics will
align with the closest octave to the reference pitch.”

"U-midics->mod12 takes in a list of list of midics and converts them to
lists of pitch class sets"

Both are document in the updated u-arithmetic patch.