R-substitute & U-divisors-draft
R-substitute

Updated to allow for different substitutions for each output voice, if
specified.  r-substitute now works like r-diminutions in this respect

Changes to r-substitute are now documented in the updated
r-substitute-ex patch.

U-divisors-draft

Draft of an object.  U-divisors-draft akes an integer as input and
returns a list of integer divisors with along with the sum of these
divisors.

Draft patch is u-divisors-draft.